Sort of sky and a couple bugfixes
I've got the rules for drawing grass all written out on paper and I've
set it up to be done. All that's left on that end is just to bunker down
and implement it. Once that's done I'll fix the rocks (which broke so
brutally they were in and out before this push) and then triangulate the
air to make it more quilt-ey